Mise à jour avec plus de fonctionnalités

+++ app/Database/Migrations/2026-08-28-130952_CreatePasswordResetsTable.php
@@ -8,21 +8,27 @@ class CreatePasswordResetsTable extends Migration
 {
     public function up()
     {
+        if ($this->db->tableExists('password_resets')) {
+            return;
+        }
+
         $this->forge->addField([
-            'id'         => ['type' => 'INT', 'constraint' => 11, 'unsigned' => true, 'auto_increment' => true],
-            'email'      => ['type' => 'VARCHAR', 'constraint' => 255],
+            'id'         => ['type' => 'INT', 'unsigned' => true, 'auto_increment' => true],
+            'email'      => ['type' => 'VARCHAR', 'constraint' => 150],
             'token'      => ['type' => 'VARCHAR', 'constraint' => 255],
             'expires_at' => ['type' => 'DATETIME'],
             'created_at' => ['type' => 'DATETIME', 'null' => true],
         ]);
+
         $this->forge->addKey('id', true);
-        $this->forge->addKey('token');
         $this->forge->addKey('email');
-        $this->forge->createTable('password_resets');
+        $this->forge->createTable('password_resets', false, ['ENGINE' => 'InnoDB']);
     }
 
     public function down()
     {
-        $this->forge->dropTable('password_resets');
+        if ($this->db->tableExists('password_resets')) {
+            $this->forge->dropTable('password_resets');
+        }
     }
 }
